Rewrite Base64Codec to prevent timing leaks
The base 64 encoder/decoder is used on secret values like encryption
keys. Unfortunately, this process is likely vulnerable to cache-timing
attacks due to the use of table lookups:

	private static final char[] CA_URL_SAFE = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_".toCharArray();

	// ...

	out[d++] = CA_URL_SAFE[(i >>> 18) & 0x3f];
	out[d++] = CA_URL_SAFE[(i >>> 12) & 0x3f];
	out[d++] = CA_URL_SAFE[(i >>> 6) & 0x3f];
	out[d++] = CA_URL_SAFE[i & 0x3f];

Here, the CA_URL_SAFE array index is a 6-bit piece of the message being
encoded. If an attacker can get information about which memory locations
are being accessed, then they may be able to figure out which array
indices were accessed, and therefore learn pieces of the
message/encryption key.